move overlay and editor controller to svgs\lines\paths
